Completed Badges for user
In this push the following changes have been included:

- AllBadgesPage. In this page I have created a class called BadgesFromCategory. In order to instantiate it, a category needs to be provided. When the category is provided a new section is built with a subtitle divider indicating the title of the category and the content itself which are the badges.

-- the content is another class called BadgesContent. There are three cases that need to be considered.

Case 1: If there are not courses in the DB for that category a text is returned to tell the user about it.

Case 2: There are courses in that category. We get a Map <courseID, courseName> for the courses in that category. To do show we call the method getCourseNamesFromCategory() of the course controller. We check the badges from the activeUser to see if the badges in his attribute collectedBadges match any of the courses in the category. If this is true we build an IconButton showing the badge. When the iconButton is pressed, an AlertDialog is displayed showing the name of the course, the picture of the badge and the time in which the user earned the badge.

Case 2.1: If the user has less than 3 badges in that category, we show greyCircles indicating where the next badges will appear.

The row of badges is composed of four circles. The first 3 are badger or grey circles (depending the case) and the last one is another IconButton +. That button when presed redirects the user to All the badges the user has earned in the category and not only three.

-CategoryBadges is the page that shows all the badges the user has earned in a certain category.