State Management
This push includes the following changes:

I have used the library GetX to introduce StateManagement on the app.

All the profile pages depend on several fields from the active user. Those fields of the user can change. For example the profile picture that the user is displaying. If those fields change the page needs to rebuild to show the changes, that is how Flutter works. There are several ways to listen to those changes and rebuild the page accordingly. One of the approaches is using stateful widgets and forcing the rebuild with the method setState(). Another approach is the Provider/consumer logic which can be used through the package provider. I have decided to use the package GetX wich offers a simple yet powerful way to handle changes. This library works in the following way:

You need to initialize a controller with all the fields of the active user. That controller extends the GetX class GetXController. I initialize this controller on the dashboard page because is where the user is first redirected when he signs-up or logs in. I initialize the controller with all the fields from the active user. You need to mark those variables as .obs which means observable because they are susceptible to changes.

Then on the different pages when we need to display any information from the active user we use the controller to get those fields. In order to use the controller and access the fields there are two ways to do it and both ways have been used in the project. In the widget that shows the info you need to extend from the GetX class known as GetView<MyController>. If you do it that way you can access the controller with the key word 'controller'. The other way is to find the active instance of the controller with the statement: Get.find().

When a widget depends on parameters from the controller that are prone to changes you need to wrap the widget with Obx(). This class lets the framework know that in case that any of the fields of the controller in which that widget depends, change then that widget needs to be rebuild.

Now I indicate for all the pages in the profile package what are the fields that are susceptible to changes and are obtained through the controller.

-Profile Page: This page shows the avatar of the activeUser. This avatar may change since we give the user the option to. This page also shows the last 4 badges earned by the user, the last 2 courses earned by the user and the last 4 avatars. In case they exist. It also shows the level of the user and its username. All the mentioned things are attributes of the user that may change thats why we get them from the controller.

-Avatars Page: This page depends on two attributes of the user, the profile picture active and the earned avatars. If one of the avatars is the one active for the user then it is shown in yellow. In this page, when the user clicks an avatar, an alert dialog is shown giving the possibility to the user to change its profile picture with that avatar. In case he decides to do so, we update the value in the controller.

- Badges Page: This page depends on the attribute Earned badges of the user. This page is composed of 4 different sections, one for each category. Each section is represented by a widget that is built with a FutureBuilder which gets the badges from a certain category. After getting those badges, I use the attribute earnedBadges from the controller to compare those badges to the ones in the category and show the ones the user has earned. Only the last 3 badges are shown. In case the user wants to see all of them he has the possibility to press the icon button provided to navigate to the page CategoryBadges where all the badges are shown. The badges are shown using an IconButton that displays a dialog with the information from the badge when pressed.

-Courses Saved: This Page is similar to the one of the badges. There are four different section, one for each category and each section is represented by a widget that is built using a FutureBuilder to get the Courses for each category. Once I get the course for a certain category, I compare them to the courses saved by the user AND the courses completed by the user. If they are the same, I display them. Only the last 3 are displayed.If the user wants to see all of the saved/completed courses of a category, he can always press a button provided to navigate to a page called CategoryCourses where all the courses and not only the last 3 are displayed. This page also includes a progress indicator in the top to show to the user how many courses he has saved and completed in that category and also how many Xp points he has earned.

Whenever the user changes one of its attributes you need to register those changes on the controller, so other widgets that depend on them are updated. Since our app has connection with a database we also need to register those changes on the user in the DB. The following things can happen: a user can save/unsave a course. A user can change its profile picture. A user can save a course for later changing the attribute currentCourse. A user can complete a course. The last action involves more changes which are the following: The level of the user may change as a consequence of completing a new course, a user may earn a new badge when he completes a new Course in case he gets more than 50% of the questions right and he has not earned that badge before and a user may earn a new Avatar if he levels up.

For each of the actions the user can take, a field or several fields in the controller are modified. To handle those changes I have created several methods on the ActiveUserController. This methods also use the UserController to persist those changes in the DB.

Previously on the code all of this methods and changes where executed through a global variable called activeUser. However, the state management requirements have made impossible to continue with this approach. Thats why this push also includes changes in all of the files that used the global variable active user to use instead the controller.